qr_generator/requirements.txt
qrcode==7.4.2
Pillow==10.1.0
pandas==2.1.3
